query syntax: add support for customizing adjust weight
We need to specify "<" or ">" to all keywords to adjust scores without
this improvement. Because the default adjust weight (6 or 4) is larger
than the default score (1). For example, "A"'s weight is 1 and "B"'s
weight is 4 in "A <B". Decremented "B"'s weight (4) is larger than not
decremented "A"'s weight (1). We need to specify ">A <B" to use
smaller weight than "A" for "B".

We just need to specify "<${WEIGHT}" or ">${WEIGHT}" only to target
keywords with this improvement. For example, "A"'s weight is 1 and
"B"'s weight is 0.9 in "A <0.1B".

There are some incompatibility but it'll be acceptable. We hope that
nobody doesn't use incompatible forms: ">${WEIGHT}...",
"<${WEIGHT}..." and "~${WEIGHT}..."

Before:

  * INCREMENT: 'A >B': "A" AND "B"
    * A's weight: 1
    * B's weight: 6

  * INCREMENT: 'A >0.1B': "A" AND "0.1B": INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * INCREMENT: 'A >"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * DECREMENT: 'A <B': "A" AND "B"
    * A's weight: 1
    * B's weight: 4

  * DECREMENT: 'A <0.1B': "A" AND "0.1B": INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 4

  * NEGATIVE: 'A ~B': "A" ("B" may be contained)
    * A's weight: 1
    * B's weight: -4

  * NEGATIVE: 'A ~0.1B': "A" ("0.1B" may be contained): INCOMPATIBLE
    * A's weight: 1
    * 0.1B's weight: -4

After:

  * INCREMENT: 'A >0.1B': "A" AND "B": INCOMPATIBLE
    * A's weight: 1
    * B's weight: 1.1

  * INCREMENT: 'A >"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 6

  * DECREMENT: 'A <0.1B': "A" AND "B": INCOMPATIBLE
    * A's weight: 1
    * B's weight: 0.9

  * DECREMENT: 'A <"0.1B"': "A" AND "0.1B": COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: 4

  * NEGATIVE: 'A ~0.1B': "A" ("B" may be contained): INCOMPATIBLE
    * A's weight: 1
    * B's weight: -0.1

  * NEGATIVE: 'A ~"0.1B"': "A" ("0.1B" may be contained): COMPATIBLE
    * A's weight: 1
    * 0.1B's weight: -4

+++ lib/grn_expr.h
@@ -86,8 +86,11 @@ scan_info **grn_scan_info_build(grn_ctx *ctx, grn_obj *expr, int *n,
 
 scan_info *grn_scan_info_open(grn_ctx *ctx, int start);
 void grn_scan_info_close(grn_ctx *ctx, scan_info *si);
-void grn_scan_info_put_index(grn_ctx *ctx, scan_info *si, grn_obj *index,
-                             uint32_t sid, int32_t weight,
+void grn_scan_info_put_index(grn_ctx *ctx,
+                             scan_info *si,
+                             grn_obj *index,
+                             uint32_t sid,
+                             float weight,
                              grn_obj *scorer,
                              grn_obj *scorer_args_expr,
                              uint32_t scorer_args_expr_offset);
@@ -116,7 +119,7 @@ int grn_scan_info_get_start_position(scan_info *si);
 void grn_scan_info_set_start_position(scan_info *si, int start);
 void grn_scan_info_reset_position(scan_info *si);
 
-int32_t grn_expr_code_get_weight(grn_ctx *ctx, grn_expr_code *ec, uint32_t *offset);
+float grn_expr_code_get_weight(grn_ctx *ctx, grn_expr_code *ec, uint32_t *offset);
 grn_rc grn_expr_code_inspect_indented(grn_ctx *ctx,
                                       grn_obj *buffer,
                                       grn_expr_code *code,
